JBPM-8766 Cannot start a task if elytron adapter is installed work aroud way
