+++ kie-server-parent/kie-server-remote/kie-server-client/pom.xml
@@ -147,6 +147,11 @@
       <artifactId>hamcrest-library</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.kie.server</groupId>
+      <artifactId>kie-server-services-common</artifactId>
+      <scope>compile</scope>
+    </dependency>
   </dependencies>
 
   <build>
