+++ kie-server-parent/kie-server-remote/kie-server-client/src/main/java/org/kie/server/client/credentials/SubjectCredentialsProvider.java
@@ -15,12 +15,8 @@
 
 package org.kie.server.client.credentials;
 
-import java.security.Principal;
-import java.util.Set;
-import javax.security.auth.Subject;
-import javax.security.jacc.PolicyContext;
-
 import org.kie.server.client.CredentialsProvider;
+import org.kie.server.services.impl.util.IdentityProviderUtils;
 
 import static javax.ws.rs.core.HttpHeaders.AUTHORIZATION;
 
@@ -34,35 +30,7 @@ public class SubjectCredentialsProvider implements CredentialsProvider {
     @Override
     public String getAuthorization() {
 
-        Subject subject = getSubjectFromContainer();
-        if (subject != null && subject.getPrincipals() != null) {
-
-            Set<Principal> principals = subject.getPrincipals();
-            for (Principal principal : principals) {
-
-                if (match(principal)) {
-                    return principal.getName();
-                }
-            }
-        }
-
-        return null;
+        return IdentityProviderUtils.getUtils().getName();
     }
 
-    // TODO make sure this can be taken out for all supported containers
-    protected Subject getSubjectFromContainer() {
-        try {
-            return (Subject) PolicyContext.getContext("javax.security.auth.Subject.container");
-        } catch (Exception e) {
-            return null;
-        }
-    }
-
-    protected boolean match(Principal principal) {
-        if (principal.getClass().getName().endsWith("BasicAuthorizationPrincipal")) {
-            return true;
-        }
-
-        return false;
-    }
 }
