+++ kie-server-parent/kie-server-services/kie-server-services-common/pom.xml
@@ -206,19 +206,23 @@
       <groupId>org.kie.soup</groupId>
       <artifactId>kie-soup-maven-support</artifactId>
     </dependency>
+    <!--<dependency>-->
+      <!--<groupId>org.keycloak</groupId>-->
+      <!--<artifactId>keycloak-core</artifactId>-->
+      <!--<exclusions>-->
+        <!--<exclusion>-->
+          <!--<groupId>org.bouncycastle</groupId>-->
+          <!--<artifactId>bcprov-jdk15on</artifactId>-->
+        <!--</exclusion>-->
+        <!--<exclusion>-->
+          <!--<groupId>org.bouncycastle</groupId>-->
+          <!--<artifactId>bcpkix-jdk15on</artifactId>-->
+        <!--</exclusion>-->
+      <!--</exclusions>-->
+    <!--</dependency>-->
     <dependency>
-      <groupId>org.keycloak</groupId>
-      <artifactId>keycloak-core</artifactId>
-      <exclusions>
-        <exclusion>
-          <groupId>org.bouncycastle</groupId>
-          <artifactId>bcprov-jdk15on</artifactId>
-        </exclusion>
-        <exclusion>
-          <groupId>org.bouncycastle</groupId>
-          <artifactId>bcpkix-jdk15on</artifactId>
-        </exclusion>
-      </exclusions>
+      <groupId>org.wildfly.security</groupId>
+      <artifactId>wildfly-elytron</artifactId>
     </dependency>
   </dependencies>
