+++ kie-server-parent/kie-server-services/kie-server-services-common/src/main/resources/META-INF/services/org.kie.server.api.security.SecurityAdapter
@@ -2,4 +2,4 @@ org.kie.server.services.impl.security.adapters.TomcatSecurityAdapter
 org.kie.server.services.impl.security.adapters.JMSSecurityAdapter
 org.kie.server.services.impl.security.adapters.WeblogicSecurityAdapter
 org.kie.server.services.impl.security.adapters.WebSphereSecurityAdapter
-org.kie.server.services.impl.security.adapters.KeycloakSecurityAdapter
+#org.kie.server.services.impl.security.adapters.KeycloakSecurityAdapter
