+++ kie-server-parent/kie-server-wars/kie-server/src/main/ee-resources/WEB-INF/web.xml
@@ -13,15 +13,15 @@
     <url-pattern>/services/rest/*</url-pattern>
   </servlet-mapping>
 
-  <filter>
-    <filter-name>request-capture</filter-name>
-    <filter-class>org.kie.server.services.impl.security.adapters.KeycloakSecurityFilter</filter-class>
-  </filter>
+  <!--<filter>-->
+    <!--<filter-name>request-capture</filter-name>-->
+    <!--<filter-class>org.kie.server.services.impl.security.adapters.KeycloakSecurityFilter</filter-class>-->
+  <!--</filter>-->
 
-  <filter-mapping>
-    <filter-name>request-capture</filter-name>
-    <url-pattern>/services/rest/*</url-pattern>
-  </filter-mapping>
+  <!--<filter-mapping>-->
+    <!--<filter-name>request-capture</filter-name>-->
+    <!--<url-pattern>/services/rest/*</url-pattern>-->
+  <!--</filter-mapping>-->
 
   <security-constraint>
     <web-resource-collection>
